feat: Introduce new dataset modules for Signature Forgery (SF) and update core scripts, DTW implementations, and documentation.

+++ dsdtw.py
@@ -82,7 +82,7 @@ class DSDTW(nn.Module):
         output = output * mask.unsqueeze(2)
 
         '''RANs'''
-        output = nutils.rnn.pack_padded_sequence(output, list(length.cpu().numpy()), batch_first=True)
+        output = nutils.rnn.pack_padded_sequence(output, list(length.cpu().to(torch.int64).numpy()), batch_first=True)
         output, hidden = self.rnn(output, self.h0)
         output, length = nutils.rnn.pad_packed_sequence(output, batch_first=True) 
         length = length.clone().detach().cuda()
